+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fix player ghosts sometimes remaining on remote clients when a different player switches a map
 - Fix remote crash when an enemy shoots any projectile
 - Fix current player floor not getting set in map menu
+- Sync ig.ENTITY.Switch position remote state
 
 ## [0.5.5] 2025-10-29